@@ -51,8 +51,8 @@ public class EnumParseTests
 		}
 	}
 
-	static readonly string[] ValidValues = new string[] { nameof(Greek.Alpha), nameof(Greek.Epsilon), nameof(Greek.Phi), nameof(Greek.Beta), nameof(Greek.Gamma) };
-	static readonly string[] InvalidValues = new string[] { "Apple", "Orange", "Pineapple", "Grapefruit", "Lemon" };
+	static readonly string[] ValidValues = [nameof(Greek.Alpha), nameof(Greek.Epsilon), nameof(Greek.Phi), nameof(Greek.Beta), nameof(Greek.Gamma)];
+	static readonly string[] InvalidValues = ["Apple", "Orange", "Pineapple", "Grapefruit", "Lemon"];
 
 	// To avoid branching overhead when benchmarking.
 	abstract class Tests
@@ -67,7 +67,7 @@ public class EnumParseTests
 
 		public abstract Greek CompiledSwitchByLength();
 
-		static readonly IDictionary<string, Greek> LookupD
+		static readonly Dictionary<string, Greek> LookupD
 			= Enum
 			.GetValues<Greek>()
 			.ToDictionary(e => Enum.GetName(e)!, e => e, StringComparer.Ordinal);
@@ -260,7 +260,7 @@ public class EnumParseTests
 			return e;
 		}
 
-		static readonly IDictionary<string, Greek> LookupD
+		static readonly Dictionary<string, Greek> LookupD
 			= Enum
 			.GetValues<Greek>()
 			.ToDictionary(e => Enum.GetName(e)!, e => e, StringComparer.OrdinalIgnoreCase);
@@ -326,7 +326,7 @@ public class EnumParseTests
 			return e;
 		}
 
-		static readonly IDictionary<string, Greek> LookupD
+		static readonly Dictionary<string, Greek> LookupD
 			= Enum
 			.GetValues<Greek>()
 			.ToDictionary(e => Enum.GetName(e)!, e => e, StringComparer.OrdinalIgnoreCase);
