@@ -1,8 +1,9 @@
 ﻿using BenchmarkDotNet.Running;
 using Open.Text.Benchmarks;
 
-BenchmarkRunner.Run<EnumParseTests>();
+//BenchmarkRunner.Run<EnumParseTests>();
 //enchmarkRunner.Run<EnumToStringTests>();
 //BenchmarkRunner.Run<CharAssumptionTests>();
 //BenchmarkRunner.Run<EnumAttributeTests>();
 //BenchmarkRunner.Run<IsDefinedTests>();
+BenchmarkRunner.Run<StringConcatTests>();
