@@ -41,6 +41,22 @@ public static partial class TextExtensions
 	}
 
 	/// <summary>
+	/// Creates a <see cref="StringSegmentEnumerable"/> for enumerating over the characters in the <paramref name="segment"/>.
+	/// </summary>
+	public static StringSegmentEnumerable AsEnumerable(this StringSegment segment)
+		=> new(segment);
+
+	/// <summary>
+	/// Creates a <see cref="CharacterSequence"/> from the provided <paramref name="sequence"/>.
+	/// </summary>
+	public static CharacterSequence AsCharacterSequence(this IReadOnlyCollection<char> sequence)
+		=> sequence is null ? CharacterSequence.Empty : CharacterSequence.Create(()=>sequence.Count, sequence!);
+
+	/// <inheritdoc cref="AsCharacterSequence(IReadOnlyCollection{char})" />
+	public static CharacterSequence AsCharacterSequence(this ICollection<char> sequence)
+		=> CharacterSequence.Create(sequence?.Count ?? 0, sequence!);
+
+	/// <summary>
 	/// Finds the first instance of a string and returns a StringSegment for subsequent use.
 	/// </summary>
 	/// <param name="source">The source string to search.</param>
@@ -310,7 +326,7 @@ public static partial class TextExtensions
 				? Enumerable.Repeat(StringSegment.Empty, 1)
 				: SplitAsSegmentsCore(source, splitCharacter),
 			StringSplitOptions.RemoveEmptyEntries => source.Length == 0
-				? Enumerable.Empty<StringSegment>()
+				? []
 				: SplitAsSegmentsCoreOmitEmpty(source, splitCharacter),
 			_ => throw new System.ComponentModel.InvalidEnumArgumentException(),
 		};
@@ -393,7 +409,7 @@ public static partial class TextExtensions
 			? throw new ArgumentNullException(nameof(pattern))
 			: source.Length == 0
 			? options == StringSplitOptions.RemoveEmptyEntries
-				? Enumerable.Empty<StringSegment>()
+				? []
 				: Enumerable.Repeat(StringSegment.Empty, 1)
 			: SplitCore(source, pattern, options);
 
@@ -448,7 +464,7 @@ public static partial class TextExtensions
 				? Enumerable.Repeat(StringSegment.Empty, 1)
 				: SplitAsSegmentsCore(source, splitSequence, comparisonType),
 			StringSplitOptions.RemoveEmptyEntries => source.Length == 0
-				? Enumerable.Empty<StringSegment>()
+				? []
 				: SplitAsSegmentsCoreOmitEmpty(source, splitSequence, comparisonType),
 			_ => throw new System.ComponentModel.InvalidEnumArgumentException(),
 		};
