@@ -10,8 +10,18 @@ namespace Open.Text;
 /// and <see cref="StringSegment.Equals(StringSegment, StringSegment, StringComparison)"/> for its comparisons
 /// and <see cref="TextExtensions.GetHashCodeFromChars(ReadOnlySpan{char}, StringComparison, int)"/> for hash codes.
 /// </summary>
-public class StringSegmentComparer
-	: IComparer<StringSegment>, IEqualityComparer<StringSegment>
+/// <remarks>
+/// Creates a new instance of <see cref="StringSegmentComparer"/>
+/// with the specified <paramref name="comparisonType"/>
+/// and <paramref name="maxHashChars"/>.
+/// </remarks>
+/// <param name="comparisonType">The string comparison type.</param>
+/// <param name="maxHashChars">
+/// The max number of characters to generate a hash from.
+/// See <see cref="TextExtensions.GetHashCodeFromChars(ReadOnlySpan{char}, StringComparison, int)"/> for more details.
+/// </param>
+public class StringSegmentComparer(StringComparison comparisonType = StringComparison.Ordinal, int maxHashChars = int.MaxValue)
+		: IComparer<StringSegment>, IEqualityComparer<StringSegment>
 {
 	/// <summary>
 	/// A <see cref="StringSegmentComparer"/> that uses <see cref="StringComparison.Ordinal"/>.
@@ -44,30 +54,14 @@ public class StringSegmentComparer
 	public static readonly StringSegmentComparer InvariantCultureIgnoreCase = new(StringComparison.InvariantCultureIgnoreCase);
 
 	/// <summary>
-	/// Creates a new instance of <see cref="StringSegmentComparer"/>
-	/// with the specified <paramref name="comparisonType"/>
-	/// and <paramref name="maxHashChars"/>.
-	/// </summary>
-	/// <param name="comparisonType">The string comparison type.</param>
-	/// <param name="maxHashChars">
-	/// The max number of characters to generate a hash from.
-	/// See <see cref="TextExtensions.GetHashCodeFromChars(ReadOnlySpan{char}, StringComparison, int)"/> for more details.
-	/// </param>
-	public StringSegmentComparer(StringComparison comparisonType = StringComparison.Ordinal, int maxHashChars = int.MaxValue)
-	{
-		ComparisonType = comparisonType;
-		MaxHashChars = maxHashChars;
-	}
-
-	/// <summary>
 	/// The string comparison type.
 	/// </summary>
-	public StringComparison ComparisonType { get; }
+	public StringComparison ComparisonType { get; } = comparisonType;
 
 	/// <summary>
 	/// The max number of characters to generate a hash from.
 	/// </summary>
-	public int MaxHashChars { get; }
+	public int MaxHashChars { get; } = maxHashChars;
 
 	/// <inheritdoc />
 	[MethodImpl(MethodImplOptions.AggressiveInlining)]
